// dac/ui/src/exports/pages/ArcticCatalog/components/ArcticCatalogDataItemSettings/ArcticTableOptimization/Optimization.types.ts
/*
 * Copyright (C) 2017-2019 Dremio Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
export type optimizationFormProps = {
  reference: string;
  tableId: string;
  catalogId: string;
  cron: string;
  setCron: (arg: string) => void;
  scheduleId?: string;
  setShowForm: (arg: boolean) => void;
  setMessage: (arg: { text: string; type: string }) => void;
  initialValues: {
    schedule: {
      scheduleType: string;
      weekInput: string[];
      monthScheduleOption: string;
      dayOfMonth: number;
      hourInput: number;
      timeInput: Date;
      ordinalOfEveryMonth: string;
      weekdayOfEveryMonth: string;
    };
    advancedConfig: {
      maxFileSize: number;
      minFileSize: number;
      minFiles: number;
      targetFileSize: number;
    };
  };
};

export type getInitialValuesResult = {
  schedule: {
    scheduleType: string;
    weekInput: string[];
    monthScheduleOption: string;
    dayOfMonth: number;
    hourInput: number;
    timeInput: Date;
    ordinalOfEveryMonth: string;
    weekdayOfEveryMonth: string;
  };
  advancedConfig: {
    maxFileSize: number;
    minFileSize: number;
    minFiles: number;
    targetFileSize: number;
  };
};

export type getInitialValuesObjectResult = {
  scheduleType: string;
  weekInput: string[];
  monthScheduleOption: string;
  dayOfMonth: number;
  ordinalOfEveryMonth: string;
  weekdayOfEveryMonth: string;
};